refactor add food
